// src/types/response/oauth.ts
export type responseTokenGoogle = {
  iss: string //'https://accounts.google.com'
  azp: string //'632048874095-hb08200odcoj9703mqcgibpcs7njfp6k.apps.googleusercontent.com'
  aud: string //'632048874095-hb08200odcoj9703mqcgibpcs7njfp6k.apps.googleusercontent.com'
  sub: string //112243752319850283693
  email: string
  email_verified: boolean
  at_hash: string //ZL031SGJJaIh_aWdWyxQpQ
  name: string
  picture: string
  given_name: string // first name manh
  family_name: string // last name quan
  locale: string // vi
  iat: number
  exp: number
}